+++ src/Api/Lookup/EntityRedirectApiLookup.php
@@ -2,8 +2,12 @@
 
 namespace Wikibase\Api\Lookup;
 
+use Mediawiki\Api\MediawikiApi;
+use Mediawiki\Api\SimpleRequest;
+use Wikibase\DataModel\Entity\BasicEntityIdParser;
 use Wikibase\DataModel\Entity\EntityId;
 use Wikibase\DataModel\Services\Lookup\EntityRedirectLookup;
+use Wikibase\DataModel\Services\Lookup\EntityRedirectLookupException;
 
 /**
  * @access private
@@ -11,34 +15,47 @@ use Wikibase\DataModel\Services\Lookup\EntityRedirectLookup;
 class EntityRedirectApiLookup implements EntityRedirectLookup {
 
 	/**
-	 * Returns the IDs that redirect to (are aliases of) the given target entity.
-	 *
-	 * @since 1.1
-	 *
-	 * @param EntityId $targetId
-	 *
-	 * @return EntityId[]
+	 * @var MediawikiApi
+	 */
+	private $api;
+
+	/**
+	 * @param MediawikiApi $api
+	 */
+	public function __construct( MediawikiApi $api ) {
+		$this->api = $api;
+	}
+
+	/**
+	 * @see EntityRedirectLookup::getRedirectIds
 	 */
 	public function getRedirectIds( EntityId $targetId ) {
 		// TODO: Implement getRedirectIds() method.
+		// Note: this is hard currently as we have to discover the namespace of the entity type?
 		throw new \BadMethodCallException('Not implemented yet');
 	}
 
 	/**
-	 * Returns the redirect target associated with the given redirect ID.
-	 *
-	 * @since 1.1
-	 *
-	 * @param EntityId $entityId
-	 * @param string $forUpdate If "for update" is given the redirect will be
-	 *        determined from the canonical master database.
-	 *
-	 * @return EntityId|null|false The ID of the redirect target, or null if $entityId
-	 *         does not refer to a redirect, or false if $entityId is not known.
+	 * @see EntityRedirectLookup::getRedirectForEntityId
 	 */
 	public function getRedirectForEntityId( EntityId $entityId, $forUpdate = '' ) {
-		// TODO: Implement getRedirectForEntityId() method.
-		throw new \BadMethodCallException('Not implemented yet');
+		$entityIdSerialization = $entityId->getSerialization();
+
+		$params = array( 'ids' => $entityIdSerialization );
+		$result = $this->api->getRequest( new SimpleRequest( 'wbgetentities', $params ) );
+
+		$entitiesData = $result['entities'];
+		if( !array_key_exists( $entityIdSerialization, $entitiesData ) ) {
+			throw new EntityRedirectLookupException( $entityId, "Failed to get $entityIdSerialization" );
+		}
+
+		$entityData = $entitiesData[$entityIdSerialization];
+		if( !array_key_exists( 'redirects', $entityData ) ) {
+			throw new EntityRedirectLookupException( $entityId, "$entityIdSerialization is not a redirect" );
+		}
+
+		$entityIdParser = new BasicEntityIdParser();
+		return $entityIdParser->parse( $entityData['redirects']['to'] );
 	}
 
 }