+++ src/Api/Service/EntityDocumentSaver.php
@@ -24,6 +24,8 @@ class EntityDocumentSaver {
 	}
 
 	/**
+	 * @since 0.7
+	 *
 	 * @param EntityDocument $entityDocument
 	 * @param EditInfo $editInfo
 	 *