+++ src/Api/WikibaseApi.php
@@ -8,6 +8,8 @@ use Mediawiki\Api\SimpleRequest;
 use Mediawiki\DataModel\EditInfo;
 
 /**
+ * @access private
+ *
  * @author Bene* < benestar.wikimedia@gmail.com >
  */
 class WikibaseApi {